requirements.txt
###### Requirements with Version Specifiers ######
python-dotenv ~= 1.0
faktory ~= 1.0
requests ~= 2.31
psycopg2-binary ~= 2.9
